re-org types
